Only push images if they have changed
We don't want to push new versions of the image if nothing has
changed. The builds aren't reproducible though, so we can't rely
on docker's layer logic to coalesce multiple builds of the same
packages.

Instead we define a `dockerdiff` that looks for differences in
the installed packages. This seems like a reasonable algorithm
for this.

We then change the push logic to first pull the currently
published image (after saving the local one to a different
name). If `dockerdiff` doesn't report any changes then
we skip pushing that image.

We control this behaviour via an env var, defaulting to off,
but we force it on if we are running in a travis cron job.
This means the daily builds will use this behaviour, but if you
push a change to this repo it will rebuild all the images. This
is because changing the scripts might change the images without
changing the package versions (e.g. changing install_packages),
so we want to trigger a push in those cases.