Tag other registries in push
Previously we were tagging for all registries as we built
the images and then pushing all registries blindly.

Now if an image hasn't changed it wouldn't tag with the other
registry tags, and then the push would fail as the gcr/quay
tags didn't exist.

Rather than taking care to tag in the case where the image
hasn't changed, instead change the pushall script to
only assume that `bitnami/minideb` tags are correct, and
tag the other registries based on that one before pushing.

This ensures that the tags will always exist, and also makes
sure we are pushing the same image to each registry (excluding
race conditions.)