Don't delete symlinks in /usr/share/doc
/usr/share/doc contains a structure that might be like

```
   foo/copyright
   foo/other
   bar -> foo
```

In this case we want to delete `foo/other`, but leave `bar` alone.

The previous code would delete `bar` as it is not a directory.

However, the structure of `/usr/share/doc` is such that the first
level is a per-package directory (or symlink) and the directory
contains the files. We therefore want to only consider files in the
package directories, so look a level further down the tree before
deleting.