Optional build with Podman (#206)
* Fix for podman usage

The output of load command in docker and podman differs:
- docker produces only one line of output the sha256 hash at 4th position
- podman produces mulktiple lines of output with sha256 hash in the last line at third position

This patch filters the line with hash and print the last argument, which is the hash. It works with docker and podman. 

Signed-off-by: flymarq <github@flynnux.de>

* Add hints for building with podman

Provide a simple command toi replace all relevant calls to docker with podman to enable a daemon less run.

Signed-off-by: flymarq <github@flynnux.de>

---------

Signed-off-by: flymarq <github@flynnux.de>