Fix keyring copy on Debian 13/trixie
It seems as though the `/usr/share/keyrings/debian-archive-keyring.gpg`
on Debian 13/trixie is actually a symlink to
`debian-archive-keyring.pgp` in the same directory. This means that the
`cp -a` in `mkimage` copies the symlink and not the actual keyring,
resulting in the following error:

    E: specified keyring file (/<tmpdir>/debian-archive-keyring.gpg) not found

Using `cp -aL` ensures the file that is symlink source actually gets
copied.

Signed-off-by: John Kristensen <john@jerrykan.com>